Title: 바구니 안의 사탕, Time: 1549ms, Memory: 61MB, Status: Wrong Answer - CodeTree

+++ 250224/바구니 안의 사탕/candy-in-the-basket.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class Main {
     public static int n,k;
     public static long min,max;
-    public static Map<Integer,Long> map = new HashMap<>();
+    public static Map<Long,Long> map = new HashMap<>();
     public static void main(String[] args) {
         // Please write your code here.
         Scanner sc = new Scanner(System.in);
